+++ pkg/driver/os_client.go
@@ -3,14 +3,14 @@ package driver
 import "os"
 
 type OsClient interface {
-	MkDirAllWithPerms(path string, perms os.FileMode, uid, gid int64) error
+	MkDirAllWithPerms(path string, perms os.FileMode, uid, gid int) error
 	Remove(path string) error
 	RemoveAll(path string) error
 }
 
 type FakeOsClient struct{}
 
-func (o *FakeOsClient) MkDirAllWithPerms(_ string, _ os.FileMode, _, _ int64) error {
+func (o *FakeOsClient) MkDirAllWithPerms(_ string, _ os.FileMode, _, _ int) error {
 	return nil
 }
 
@@ -24,7 +24,7 @@ func (o *FakeOsClient) RemoveAll(_ string) error {
 
 type BrokenOsClient struct{}
 
-func (o *BrokenOsClient) MkDirAllWithPerms(_ string, _ os.FileMode, _, _ int64) error {
+func (o *BrokenOsClient) MkDirAllWithPerms(_ string, _ os.FileMode, _, _ int) error {
 	return &os.PathError{}
 }
 
@@ -38,12 +38,12 @@ func (o *BrokenOsClient) RemoveAll(_ string) error {
 
 type RealOsClient struct{}
 
-func (o *RealOsClient) MkDirAllWithPerms(path string, perms os.FileMode, uid, gid int64) error {
+func (o *RealOsClient) MkDirAllWithPerms(path string, perms os.FileMode, uid, gid int) error {
 	err := os.MkdirAll(path, perms)
 	if err != nil {
 		return err
 	}
-	err = os.Chown(path, int(uid), int(gid))
+	err = os.Chown(path, uid, gid)
 	if err != nil {
 		return err
 	}
